Change serialization API.
To implement serialization for a user-defined type,
tpie::serialize/tpie::unserialize must be overridden (in the tpie
namespace). This is similar to the std::swap API, for which the C++
standard allows template specialization in the std namespace. (N3242
17.6.4.2.1 paragraph 1.)

This changeset removes the tp_serialize and tp_unserialize functions.

Adds a test that ensures tp_serialize can be overridden in user code.